+++ eslintrc-ts.yml
@@ -1,0 +1,5 @@
+extends: ['./.eslintrc.yml', 'plugin:@typescript-eslint/recommended']
+parser: '@typescript-eslint/parser'
+rules:
+  '@typescript-eslint/explicit-function-return-type': 'off'
+  '@typescript-eslint/no-use-before-define': 'off'
