+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "build": "webpack --mode production",
     "start": "webpack-dev-server",
-    "eslint": "eslint --ext .jsx --ext .js src/",
+    "eslint.js": "eslint --ext .jsx --ext .js src/",
+    "eslint.ts": "eslint --config .eslintrc-tc.yml --ext .tsx src/",
     "test": "npm run eslint && npm run test.jest",
     "test.jest": "jest"
   },
@@ -14,6 +15,11 @@
     "@babel/core": "7.10.2",
     "@babel/preset-env": "7.10.2",
     "@babel/preset-react": "7.10.1",
+    "@types/jest": "26.0.8",
+    "@types/react": "16.9.44",
+    "@types/react-test-renderer": "16.9.2",
+    "@typescript-eslint/eslint-plugin": "3.7.1",
+    "@typescript-eslint/parser": "3.7.1",
     "axios-mock-adapter": "^1.18.1",
     "babel-jest": "26.0.1",
     "babel-loader": "8.1.0",
@@ -25,6 +31,10 @@
     "jest": "26.0.1",
     "react-test-renderer": "16.13.1",
     "redux-mock-store": "1.5.4",
+    "ts-jest": "26.1.4",
+    "ts-loader": "8.0.2",
+    "typescript": "3.9.7",
+    "utility-types": "3.10.0",
     "webpack": "4.43.0",
     "webpack-cli": "3.3.11",
     "webpack-dev-server": "3.11.0"
