+++ src/components/comments/comments.tsx
@@ -1,6 +1,6 @@
-import React from "react";
+import * as React from "react"
 import PropTypes from 'prop-types';
-import {commentInputs} from '../../constants.js';
+import {commentInputs} from '../../constants';
 
 class Comments extends React.PureComponent {
   constructor(props) {
