+++ src/components/empty/empty.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import * as React from "react"
 
 const Empty = () => {
   return (
