+++ src/components/private-route/private-route.tsx
@@ -1,8 +1,8 @@
-import React from "react";
+import * as React from "react"
 import PropTypes from "prop-types";
 import {Route, Redirect} from "react-router-dom";
-import {AppRoute} from "../../constants.js";
-import {userType} from "../../types/dataTypes.js";
+import {AppRoute} from "../../constants";
+import {userType} from "../../types/dataTypes";
 
 const PrivateRoute = ({component: Component, ...rest}) => {
   return <Route
