+++ src/hocs/with-sort/with-sort.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import * as React from "react"
 
 const sortTypeMapping = [
   {
