+++ webpack.config.js
@@ -1,15 +1,22 @@
 const path = require(`path`);
 
 module.exports = {
-  entry: `./src/index.js`,
+  entry: `./src/index.tsx`,
   output: {
     filename: `bundle.js`,
     path: path.join(__dirname, `public`)
   },
   devServer: {
     contentBase: path.join(__dirname, `public`),
-    open: false,
+    open: true,
     port: 1337,
+    publicPath: `/`,
+    proxy: {
+      '/offer': {
+        target: `http://localhost:1337/`,
+        pathRewrite: {'^/offer': ``},
+      },
+    },
     historyApiFallback: true
   },
   module: {
@@ -20,8 +27,15 @@ module.exports = {
         use: {
           loader: `babel-loader`,
         },
+      },
+      {
+        test: /\.(tsx|ts)?$/,
+        loader: `ts-loader`
       }
     ],
+  },
+  resolve: {
+    extensions: [`.ts`, `.tsx`, `.js`, `json`]
   },
   devtool: `source-map`
 };
